LayoutTests/css3/flexbox/flex-shorthand-flex-basis-middle.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox {
    width: 500px;
    height:100px;
    background: green;
}

.flexbox > div {
    width: 50px;
    background: yellow;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>
<div class="flexbox">
    <div data-expected-width="50" style="flex: 1 0% 1"></div>
    <div data-expected-width="50" style="flex: 1 0% 1"></div>
    <div data-expected-width="50" style="flex: 1 0% 1"></div>
</div>

<div class="flexbox">
    <div data-expected-width="50" style="flex: 1 auto 1"></div>
    <div data-expected-width="50" style="flex: 1 auto 1"></div>
    <div data-expected-width="50" style="flex: 1 auto 1"></div>
</div>
</body>
</html>
